+++ deepchem/__init__.py
@@ -14,4 +14,3 @@ import deepchem.nn
 import deepchem.splits
 import deepchem.trans
 import deepchem.utils
-import deepchem.load
