+++ deepchem/data/__init__.py
@@ -8,4 +8,5 @@ from __future__ import unicode_literals
 # TODO(rbharath): Get rid of * import
 from deepchem.data.datasets import *
 from deepchem.data.supports import *
+from deepchem.data.data_loader import DataLoader
 import deepchem.data.tests
