+++ deepchem/data/tests/__init__.py
@@ -23,11 +23,8 @@ def load_solubility_data():
   tasks = ["log-solubility"]
   task_type = "regression"
   input_file = os.path.join(current_dir, "../../models/tests/example.csv")
-  featurizer = dc.load.DataLoader(
-      tasks=tasks,
-      smiles_field="smiles",
-      featurizer=featurizer,
-      verbosity="low")
+  featurizer = dc.data.DataLoader(
+      tasks=tasks, smiles_field="smiles", featurizer=featurizer)
   return featurizer.featurize(input_file)
 
 def load_multitask_data():
@@ -39,11 +36,8 @@ def load_multitask_data():
            "task13", "task14", "task15", "task16"]
   input_file = os.path.join(
       current_dir, "../../models/tests/multitask_example.csv")
-  loader = dc.load.DataLoader(
-      tasks=tasks,
-      smiles_field="smiles",
-      featurizer=featurizer,
-      verbosity="low")
+  loader = dc.data.DataLoader(
+      tasks=tasks, smiles_field="smiles", featurizer=featurizer)
   return loader.featurize(input_file)
 
 def load_classification_data():
@@ -54,9 +48,8 @@ def load_classification_data():
   task_type = "classification"
   input_file = os.path.join(
       current_dir, "../../models/tests/example_classification.csv")
-  loader = dc.load.DataLoader(
-      tasks=tasks, smiles_field="smiles",
-      featurizer=featurizer, verbosity="low")
+  loader = dc.data.DataLoader(
+      tasks=tasks, smiles_field="smiles", featurizer=featurizer)
   return loader.featurize(input_file)
 
 
@@ -68,9 +61,8 @@ def load_sparse_multitask_dataset():
            "task7", "task8", "task9"]
   input_file = os.path.join(
       current_dir, "../../models/tests/sparse_multitask_example.csv")
-  loader = dc.load.DataLoader(
-      tasks=tasks, smiles_field="smiles",
-      featurizer=featurizer, verbosity="low")
+  loader = dc.data.DataLoader(
+      tasks=tasks, smiles_field="smiles", featurizer=featurizer)
   return loader.featurize(input_file)
   
 def load_feat_multitask_data():
@@ -81,9 +73,8 @@ def load_feat_multitask_data():
   tasks = ["task0", "task1", "task2", "task3", "task4", "task5"]
   input_file = os.path.join(
       current_dir, "../../models/tests/feat_multitask_example.csv")
-  loader = dc.load.DataLoader(
-      tasks=tasks, featurizer=featurizer,
-      id_field="id", verbosity="low")
+  loader = dc.data.DataLoader(
+      tasks=tasks, featurizer=featurizer, id_field="id")
   return loader.featurize(input_file)
 
 def load_gaussian_cdf_data():
@@ -96,7 +87,6 @@ def load_gaussian_cdf_data():
   tasks = ["task0","task1"]
   input_file = os.path.join(
       current_dir, "../../models/tests/gaussian_cdf_example.csv")
-  loader = dc.load.DataLoader(
-      tasks=tasks, featurizer=featurizer,
-      id_field="id", verbosity=None)
+  loader = dc.data.DataLoader(
+      tasks=tasks, featurizer=featurizer, id_field="id")
   return loader.featurize(input_file)
