+++ deepchem/data/tests/test_load.py
@@ -23,7 +23,6 @@ class TestLoad(unittest.TestCase):
 
   def test_move_load(self):
     """Test that datasets can be moved and loaded."""
-    verbosity = "high"
     current_dir = os.path.dirname(os.path.realpath(__file__))
     base_dir = tempfile.mkdtemp()
     data_dir = os.path.join(base_dir, "data")
@@ -33,9 +32,8 @@ class TestLoad(unittest.TestCase):
 
     featurizer = dc.feat.CircularFingerprint(size=1024)
     tasks = ["log-solubility"]
-    loader = dc.load.DataLoader(
-        tasks=tasks, smiles_field="smiles",
-        featurizer=featurizer, verbosity=verbosity)
+    loader = dc.data.DataLoader(
+        tasks=tasks, smiles_field="smiles", featurizer=featurizer)
     dataset = loader.featurize(dataset_file, data_dir)
 
     X, y, w, ids = (dataset.X, dataset.y, dataset.w, dataset.ids)
@@ -60,7 +58,6 @@ class TestLoad(unittest.TestCase):
 
     # Set some global variables up top
     reload = True
-    verbosity = "high"
 
     current_dir = os.path.dirname(os.path.realpath(__file__))
     ##Make directories to store the raw and featurized datasets.
@@ -77,9 +74,8 @@ class TestLoad(unittest.TestCase):
     all_tasks = ["task%d"%i for i in range(17)] 
 
     ####### Do featurization
-    loader = dc.load.DataLoader(
-        tasks=all_tasks, smiles_field="smiles",
-        featurizer=featurizer, verbosity=verbosity)
+    loader = dc.data.DataLoader(
+        tasks=all_tasks, smiles_field="smiles", featurizer=featurizer)
     dataset = loader.featurize(dataset_file, data_dir)
 
     # Do train/valid split.
@@ -91,8 +87,7 @@ class TestLoad(unittest.TestCase):
     y_tasks, w_tasks, = [], []
     for ind, task in enumerate(all_tasks):
       print("Processing task %s" % task)
-      dataset = dc.data.DiskDataset(data_dir, verbosity=verbosity,
-                                        reload=reload)
+      dataset = dc.data.DiskDataset(data_dir, reload=reload)
 
       X_task, y_task, w_task, ids_task = (dataset.X, dataset.y, dataset.w,
                                           dataset.ids)
@@ -117,7 +112,6 @@ class TestLoad(unittest.TestCase):
 
     # Set some global variables up top
     reload = True
-    verbosity = "high"
 
     current_dir = os.path.dirname(os.path.realpath(__file__))
     #Make directories to store the raw and featurized datasets.
@@ -137,9 +131,8 @@ class TestLoad(unittest.TestCase):
     tasks = all_tasks[0:n_tasks]
 
     ####### Do multitask load
-    loader = dc.load.DataLoader(
-        tasks=tasks, smiles_field="smiles",
-        featurizer=featurizer, verbosity=verbosity)
+    loader = dc.data.DataLoader(
+        tasks=tasks, smiles_field="smiles", featurizer=featurizer)
     dataset = loader.featurize(dataset_file, data_dir)
 
     # Do train/valid split.
@@ -152,9 +145,8 @@ class TestLoad(unittest.TestCase):
       print("Processing task %s" % task)
       if os.path.exists(data_dir):
         shutil.rmtree(data_dir)
-      loader = dc.load.DataLoader(
-          tasks=[task], smiles_field="smiles",
-          featurizer=featurizer, verbosity=verbosity)
+      loader = dc.data.DataLoader(
+          tasks=[task], smiles_field="smiles", featurizer=featurizer)
       dataset = loader.featurize(dataset_file, data_dir)
 
       X_task, y_task, w_task, ids_task = (dataset.X, dataset.y, dataset.w,
