+++ deepchem/hyper/tests/test_hyperparam_opt.py
@@ -29,9 +29,8 @@ class TestHyperparamOptAPI(unittest.TestCase):
     tasks = ["log-solubility"]
     current_dir = os.path.dirname(os.path.abspath(__file__))
     input_file = os.path.join(current_dir, "../../models/tests/example.csv")
-    loader = dc.load.DataLoader(
-        tasks=tasks, smiles_field="smiles",
-        featurizer=featurizer, verbosity="low")
+    loader = dc.data.DataLoader(
+        tasks=tasks, smiles_field="smiles", featurizer=featurizer)
     dataset = loader.featurize(input_file)
 
     splitter = dc.splits.ScaffoldSplitter()
@@ -113,9 +112,8 @@ class TestHyperparamOptAPI(unittest.TestCase):
     n_features = 1024
     featurizer = dc.feat.CircularFingerprint(size=n_features)
 
-    loader = dc.load.DataLoader(
-        tasks=tasks, smiles_field="smiles",
-        featurizer=featurizer, verbosity="low")
+    loader = dc.data.DataLoader(
+        tasks=tasks, smiles_field="smiles", featurizer=featurizer)
     dataset = loader.featurize(input_file)
 
     splitter = dc.splits.ScaffoldSplitter()
