--- deepchem/load/__init__.py
@@ -1,12 +1,0 @@
-"""
-Wrapper for deepchem data inputs.
-"""
-from __future__ import print_function
-from __future__ import division
-from __future__ import unicode_literals
-
-__author__ = "Bharath Ramsundar"
-__copyright__ = "Copyright 2016, Stanford University"
-__license__ = "GPL"
-
-from deepchem.feat.featurize import DataLoader
