+++ deepchem/models/tests/test_api.py
@@ -28,9 +28,9 @@ class TestAPI(unittest.TestCase):
     tasks = ["log-solubility"]
     current_dir = os.path.dirname(os.path.abspath(__file__))
     input_file = os.path.join(current_dir, "example.csv")
-    loader = dc.load.DataLoader(
+    loader = dc.data.DataLoader(
         tasks=tasks, smiles_field="smiles",
-        featurizer=featurizer, verbosity="low")
+        featurizer=featurizer)
     dataset = loader.featurize(input_file)
 
     splitter = dc.splits.ScaffoldSplitter()
@@ -61,9 +61,8 @@ class TestAPI(unittest.TestCase):
     tasks = ["log-solubility"]
     current_dir = os.path.dirname(os.path.abspath(__file__))
     input_file = os.path.join(current_dir, "user_specified_example.csv")
-    loader = dc.load.DataLoader(
-        tasks=tasks, smiles_field="smiles", featurizer=featurizer,
-        verbosity="low")
+    loader = dc.data.DataLoader(
+        tasks=tasks, smiles_field="smiles", featurizer=featurizer)
     dataset = loader.featurize(input_file, debug=True)
 
     splitter = dc.splits.SpecifiedSplitter(input_file, "split")
@@ -98,10 +97,9 @@ class TestAPI(unittest.TestCase):
 
     current_dir = os.path.dirname(os.path.abspath(__file__))
     input_file = os.path.join(current_dir, "example.csv")
-    loader = dc.load.DataLoader(tasks=tasks,
+    loader = dc.data.DataLoader(tasks=tasks,
                         smiles_field="smiles",
-                        featurizer=featurizer,
-                        verbosity="low")
+                        featurizer=featurizer)
     dataset = loader.featurize(input_file)
 
     splitter = dc.splits.ScaffoldSplitter()
@@ -142,9 +140,9 @@ class TestAPI(unittest.TestCase):
     current_dir = os.path.dirname(os.path.abspath(__file__))
     input_file = os.path.join(current_dir, "example_classification.csv")
 
-    loader = dc.load.DataLoader(
+    loader = dc.data.DataLoader(
         tasks=tasks, smiles_field="smiles",
-        featurizer=featurizer, verbosity="low")
+        featurizer=featurizer)
     dataset = loader.featurize(input_file)
 
     splitter = dc.splits.ScaffoldSplitter()
